more review comments and unit test fixes

+++ indexing-service/src/test/java/io/druid/indexing/common/task/AppenderatorDriverRealtimeIndexTaskTest.java
@@ -223,7 +223,8 @@ public class AppenderatorDriverRealtimeIndexTaskTest
     @Override
     public Runnable commit()
     {
-      return () -> {};
+      return () -> {
+      };
     }
 
     @Override
@@ -330,7 +331,8 @@ public class AppenderatorDriverRealtimeIndexTaskTest
 
     // handoff would timeout, resulting in exception
     TaskStatus status = statusFuture.get();
-    Assert.assertTrue(status.getErrorMsg().contains("java.util.concurrent.TimeoutException: Timeout waiting for task."));
+    Assert.assertTrue(status.getErrorMsg()
+                            .contains("java.util.concurrent.TimeoutException: Timeout waiting for task."));
   }
 
   @Test(timeout = 60_000L)
@@ -366,8 +368,8 @@ public class AppenderatorDriverRealtimeIndexTaskTest
     Assert.assertEquals(0, task.getMetrics().unparseable());
 
     // Do some queries.
-    Assert.assertEquals(2, sumMetric(task, null, "rows"));
-    Assert.assertEquals(3, sumMetric(task, null, "met1"));
+    Assert.assertEquals(2, sumMetric(task, null, "rows").longValue());
+    Assert.assertEquals(3, sumMetric(task, null, "met1").longValue());
 
     awaitHandoffs();
 
@@ -428,8 +430,8 @@ public class AppenderatorDriverRealtimeIndexTaskTest
     Assert.assertEquals(0, task.getMetrics().unparseable());
 
     // Do some queries.
-    Assert.assertEquals(2, sumMetric(task, null, "rows"));
-    Assert.assertEquals(3, sumMetric(task, null, "met1"));
+    Assert.assertEquals(2, sumMetric(task, null, "rows").longValue());
+    Assert.assertEquals(3, sumMetric(task, null, "met1").longValue());
 
     awaitHandoffs();
 
@@ -493,8 +495,8 @@ public class AppenderatorDriverRealtimeIndexTaskTest
     Assert.assertEquals(0, task.getMetrics().unparseable());
 
     // Do some queries.
-    Assert.assertEquals(2000, sumMetric(task, null, "met1"));
-    Assert.assertEquals(2000, sumMetric(task, null, "met1"));
+    Assert.assertEquals(2000, sumMetric(task, null, "met1").longValue());
+    Assert.assertEquals(2000, sumMetric(task, null, "met1").longValue());
 
     awaitHandoffs();
 
@@ -561,14 +563,14 @@ public class AppenderatorDriverRealtimeIndexTaskTest
     Assert.assertEquals(0, task.getMetrics().unparseable());
 
     // Do some queries.
-    Assert.assertEquals(2, sumMetric(task, null, "rows"));
-    Assert.assertEquals(2, sumMetric(task, new SelectorDimFilter("dim1t", "foofoo", null), "rows"));
+    Assert.assertEquals(2, sumMetric(task, null, "rows").longValue());
+    Assert.assertEquals(2, sumMetric(task, new SelectorDimFilter("dim1t", "foofoo", null), "rows").longValue());
     if (NullHandling.replaceWithDefault()) {
-      Assert.assertEquals(0, sumMetric(task, new SelectorDimFilter("dim1t", "barbar", null), "metric1"));
+      Assert.assertEquals(0, sumMetric(task, new SelectorDimFilter("dim1t", "barbar", null), "metric1").longValue());
     } else {
       Assert.assertNull(sumMetric(task, new SelectorDimFilter("dim1t", "barbar", null), "metric1"));
     }
-    Assert.assertEquals(3, sumMetric(task, null, "met1"));
+    Assert.assertEquals(3, sumMetric(task, null, "met1").longValue());
 
     awaitHandoffs();
 
@@ -623,7 +625,8 @@ public class AppenderatorDriverRealtimeIndexTaskTest
 
     // Wait for the task to finish.
     TaskStatus status = statusFuture.get();
-    Assert.assertTrue(status.getErrorMsg().contains("java.lang.RuntimeException: Max parse exceptions exceeded, terminating task..."));
+    Assert.assertTrue(status.getErrorMsg()
+                            .contains("java.lang.RuntimeException: Max parse exceptions exceeded, terminating task..."));
 
     IngestionStatsAndErrorsTaskReportData reportData = getTaskReportData();
 
@@ -642,7 +645,15 @@ public class AppenderatorDriverRealtimeIndexTaskTest
   {
     expectPublishedSegments(1);
 
-    final AppenderatorDriverRealtimeIndexTask task = makeRealtimeTask(null, TransformSpec.NONE, false, 0, true, null, 1);
+    final AppenderatorDriverRealtimeIndexTask task = makeRealtimeTask(
+        null,
+        TransformSpec.NONE,
+        false,
+        0,
+        true,
+        null,
+        1
+    );
     final ListenableFuture<TaskStatus> statusFuture = runTask(task);
 
     // Wait for firehose to show up, it starts off null.
@@ -686,8 +697,8 @@ public class AppenderatorDriverRealtimeIndexTaskTest
     Assert.assertEquals(2, task.getMetrics().unparseable());
 
     // Do some queries.
-    Assert.assertEquals(3, sumMetric(task, null, "met1"));
-    Assert.assertEquals(3, sumMetric(task, null, "met1"));
+    Assert.assertEquals(3, sumMetric(task, null, "met1").longValue());
+    Assert.assertEquals(3, sumMetric(task, null, "met1").longValue());
 
     awaitHandoffs();
 
@@ -752,7 +763,18 @@ public class AppenderatorDriverRealtimeIndexTaskTest
             ImmutableMap.of("t", 1521251960729L, "dim1", "foo", "met1", "foo"),
 
             // Bad long dim- will count as processed, but bad dims will get default values
-            ImmutableMap.of("t", 1521251960729L, "dim1", "foo", "dimLong", "notnumber", "dimFloat", "notnumber", "met1", "foo"),
+            ImmutableMap.of(
+                "t",
+                1521251960729L,
+                "dim1",
+                "foo",
+                "dimLong",
+                "notnumber",
+                "dimFloat",
+                "notnumber",
+                "met1",
+                "foo"
+            ),
 
             // Bad row- will be unparseable.
             ImmutableMap.of("dim1", "foo", "met1", 2.0, FAIL_DIM, "x"),
@@ -777,8 +799,8 @@ public class AppenderatorDriverRealtimeIndexTaskTest
     Assert.assertEquals(2, task.getMetrics().unparseable());
 
     // Do some queries.
-    Assert.assertEquals(4, sumMetric(task, null, "rows"));
-    Assert.assertEquals(3, sumMetric(task, null, "met1"));
+    Assert.assertEquals(4, sumMetric(task, null, "rows").longValue());
+    Assert.assertEquals(3, sumMetric(task, null, "met1").longValue());
 
     awaitHandoffs();
 
@@ -854,7 +876,18 @@ public class AppenderatorDriverRealtimeIndexTaskTest
             ImmutableMap.of("t", 1521251960729L, "dim1", "foo", "met1", "foo"),
 
             // Bad long dim- will count as processed, but bad dims will get default values
-            ImmutableMap.of("t", 1521251960729L, "dim1", "foo", "dimLong", "notnumber", "dimFloat", "notnumber", "met1", "foo"),
+            ImmutableMap.of(
+                "t",
+                1521251960729L,
+                "dim1",
+                "foo",
+                "dimLong",
+                "notnumber",
+                "dimFloat",
+                "notnumber",
+                "met1",
+                "foo"
+            ),
 
             // Bad row- will be unparseable.
             ImmutableMap.of("dim1", "foo", "met1", 2.0, FAIL_DIM, "x"),
@@ -945,7 +978,7 @@ public class AppenderatorDriverRealtimeIndexTaskTest
       }
 
       // Do a query, at this point the previous data should be loaded.
-      Assert.assertEquals(1, sumMetric(task2, null, "rows"));
+      Assert.assertEquals(1, sumMetric(task2, null, "rows").longValue());
 
       final TestFirehose firehose = (TestFirehose) task2.getFirehose();
 
@@ -963,7 +996,7 @@ public class AppenderatorDriverRealtimeIndexTaskTest
       publishedSegment = Iterables.getOnlyElement(publishedSegments);
 
       // Do a query.
-      Assert.assertEquals(2, sumMetric(task2, null, "rows"));
+      Assert.assertEquals(2, sumMetric(task2, null, "rows").longValue());
 
       awaitHandoffs();
 
@@ -1020,7 +1053,7 @@ public class AppenderatorDriverRealtimeIndexTaskTest
       publishedSegment = Iterables.getOnlyElement(publishedSegments);
 
       // Do a query.
-      Assert.assertEquals(1, sumMetric(task1, null, "rows"));
+      Assert.assertEquals(1, sumMetric(task1, null, "rows").longValue());
 
       // Trigger graceful shutdown.
       task1.stopGracefully();
@@ -1139,7 +1172,8 @@ public class AppenderatorDriverRealtimeIndexTaskTest
           )
       );
       Assert.assertEquals(expectedMetrics, reportData.getRowStats());
-      Assert.assertTrue(status.getErrorMsg().contains("java.lang.IllegalArgumentException\n\tat java.nio.Buffer.position"));
+      Assert.assertTrue(status.getErrorMsg()
+                              .contains("java.lang.IllegalArgumentException\n\tat java.nio.Buffer.position"));
     }
   }
 
@@ -1465,7 +1499,7 @@ public class AppenderatorDriverRealtimeIndexTaskTest
     );
   }
 
-  public long sumMetric(final Task task, final DimFilter filter, final String metric)
+  public Long sumMetric(final Task task, final DimFilter filter, final String metric)
   {
     // Do a query.
     TimeseriesQuery query = Druids.newTimeseriesQueryBuilder()
